docs: update on java files

+++ exercicios/ExercicioTres.java
@@ -4,10 +4,13 @@ import java.util.Scanner;
 public class ExercicioTres {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
+
         System.out.println("| Conversor de temperatura |");
         System.out.println("Digite a temperatura em Fahrenheit");
-        int temperatureinFahrenheit = scan.nextInt();
-        int temperatureinCelsius = 5*((temperatureinFahrenheit-32)/9);
+        double temperatureinFahrenheit = scan.nextDouble();
+        
+        double temperatureinCelsius = 5*((temperatureinFahrenheit-32)/9);
+        
         System.out.println(temperatureinFahrenheit + " em graus Celsius eh: " + temperatureinCelsius + ".\n"); 
         scan.close();
     }
